#REFACTOR Templates uebernehmen Bestimmung der Spaltenanzahl
